adds agentInstallationStage=-1,0 handling
